requirements.txt
datasets 
pyyaml
transformers
peft
accelerate